# openapi.yaml
openapi: 3.1.0
info:
  title: GABAI API
  version: 0.1.0
  description: |
    Citizen service request classification and routing for a barangay LGU.

    Residents submit free-text requests; a classifier assigns a category and an
    urgency level, each with its own confidence score. A request is auto-routed
    only when `min(category_confidence, urgency_confidence)` meets the configured
    threshold; otherwise it enters the manual review queue for a human to label.

    **Conventions**

    - JSON bodies use `snake_case`.
    - Timestamps are ISO-8601 with a UTC offset (`2026-08-30T04:15:00Z`).
    - List endpoints use offset pagination and return a `PaginatedResponse`.
    - Errors use a single envelope; see `Error`.
    - Authorization is enforced server-side on every endpoint. A client hiding a
      control is not access control.
    - Categories are **data, not an enum**. Never hardcode them; read
      `GET /api/categories`.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: /
    description: Same origin as the deployed frontend
...
security:
  - cookieAuth: []
paths:
...
  /api/admin/users:
    get:
      tags: [admin]
      summary: List accounts
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/Role"
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          description: Case-insensitive substring match on email and full name.
          schema:
            type: string
            maxLength: 200
      responses:
        "200":
          description: A page of accounts.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [admin]
      summary: Create an account with a role
      description: The only way a staff or admin account comes into existence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/admin/users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [admin]
      summary: Update an account
      description: |
        Accounts are deactivated, never deleted. Deleting one would orphan the
        audit trail and the requests it acted on. Set `is_active: false` instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/admin/users/{user_id}/password:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [admin]
      summary: Reset an account password
      description: |
        Issues a temporary password and returns it once, in this response. It is
        never stored unhashed and never appears in the audit log, which any admin
        can read. The admin reads it to the account holder, who changes it.

        Writes an audit row with action `user.password_reset` and no detail.
        Sets `must_change_password`. The temporary password can still sign in,
        but that session can only change the password until it's replaced.

        An admin may reset their own account, so there is no 409 here.
      responses:
        "200":
          description: Password reset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemporaryPassword"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gabai_session
      description: |
        HttpOnly session cookie set by `POST /api/auth/login`. Browsers attach it
        automatically; send credentialed requests (`credentials: "include"` on
        cross-origin fetches).
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    Unauthorized:
      description: No session, or the session has expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated, but not allowed. The role does not permit this action,
        the account is still waiting for approval, it is on a temporary
        password and has to change it first, or it hasn't accepted the current
        Terms of Use and Privacy Notice.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        No such resource, or it exists outside the caller's scope. The two are
        deliberately indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflicts with existing state, e.g. the email is taken.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: The body failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier. Safe to branch on.
              examples: [not_found, forbidden, validation_error, invalid_transition]
            message:
              type: string
              description: Human-readable, safe to show a barangay staff member.
            details:
              type: array
              description: Per-field problems, present on validation errors.
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    examples: [description]
                  message:
                    type: string
                    examples: ["Field is required."]
    PaginatedResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total matching rows, ignoring limit and offset.
        limit:
          type: integer
        offset:
          type: integer
...
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected]
      description: Staff and admin accounts are always `approved`.
    Role:
      type: string
      enum: [citizen, staff, admin]
...
    User:
      type: object
      required: [id, email, full_name, role, is_active, is_verified, approval_status, must_change_password, terms_accepted_version, terms_accepted_at, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: |
            UUID rather than a sequential integer. This is the only table holding
            personal data and its ids appear in URLs and audit records, so an
            authorization slip must not degrade into enumerating every resident.
        email:
          type: string
          format: email
        full_name:
          type: string
          maxLength: 150
        role:
          $ref: "#/components/schemas/Role"
        is_active:
          type: boolean
        is_verified:
          type: boolean
          description: The person clicked the link emailed to this address.
        approval_status:
          $ref: "#/components/schemas/ApprovalStatus"
        residence:
          type: [string, "null"]
          description: |
            Purok or street, as the resident typed it at sign-up. Only there so
            staff can confirm the person lives in the barangay. Null for staff
            and admin accounts.
        must_change_password:
          type: boolean
          description: |
            Set when an admin creates the account or resets its password, cleared
            when the person sets their own. Enforced server-side.
        terms_accepted_version:
          type: [string, "null"]
          description: |
            The Terms of Use and Privacy Notice version this person last accepted.
            Null for accounts made before acceptance was recorded. When it differs
            from `terms_version` in `GET /api/site`, the client asks them to accept.
        terms_accepted_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
...
    UserCreate:
      description: Admin-created accounts are approved from the start.
      type: object
      required: [email, password, full_name, role]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        full_name:
          type: string
          minLength: 1
          maxLength: 150
        role:
          $ref: "#/components/schemas/Role"
    UserUpdate:
      type: object
      description: Only the supplied fields change. A name is trimmed, and a change of name is recorded in the audit log as `user.renamed`, without the names.
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 150
        role:
          $ref: "#/components/schemas/Role"
        is_active:
          type: boolean
...
    TemporaryPassword:
      type: object
      description: |
        Returned once and never again. The plaintext exists only in this response
        body; the stored value is hashed like any other password.
      required: [temporary_password]
      properties:
        temporary_password:
          type: string
          description: |
            Ten characters from an alphabet with `0`, `O`, `1`, `l` and `I`
            removed, because someone reads this aloud at a shared desk.

    # ---------------------------------------------------- categories ------
...